Implement the search field
